requirements.txt
# this is the "requirements.txt" file (top level)
# we list the name of each package this app requires

pandas
requests
plotly
python-dotenv
sendgrid==6.5.0